+++ mvc/src/main/java/org/karnak/data/AppConfig.java
@@ -8,7 +8,6 @@ import org.karnak.profilepipe.Profiles;
 import org.karnak.profilepipe.profilebody.ProfilePipeBody;
 import org.karnak.standard.ConfidentialityProfiles;
 import org.karnak.standard.StandardDICOM;
-import org.karnak.ui.extid.Patient;
 import org.karnak.ui.profile.ProfilePipeService;
 import org.karnak.ui.profile.ProfilePipeServiceImpl;
 import org.slf4j.Logger;
